[TASK] Restructure "Publish an Extension"
The general recommendations on "Publish on TER" are not specific to publishing
on TER. It is recommended to follow all the publishing steps but it is also
possible (and common practice) to follow only a subset, e.g. publish only on
Packagist.

A previous change added a number of recommendations to the page about
Publishing to TER which are general and not specific to TER. Also
pointing out the only real requirement (and not recommendation),
the extension key (and the rules for it), was missing.

This patch restructures the pages slightly, adding a general
Prerequisites page and adding links to it where appropriate.

The order is changed, so that a link to the Prerequisites page is
on the top of the start page.

Also, where a rule / recommendation is a requirement, this information
is added as "requirement" or using terms such as MUST etc.

The additional metadata on TER is also described (not just the link
to issues and repo), but also documentation because that is a common
"gotcha" to add a link to the documentation where it is not necessary.